requirements.txt
discord.py>=2.5,<3.0
python-dotenv>=1.0,<2.0
aiohttp>=3.9,<4.0
